refactor(guardrails): extract pr-terminal types and constants to module boundary
Review follow-up: PrTerminalVerdict/PrTerminalEvent/PrTerminalGateAction/
PrTerminalGateDecision move to types.ts, the detection regexes and
MAX_PR_TERMINAL_BLOCKS to a new constants.ts (rationale comments travel
with the regexes); prTerminal.ts keeps only the four functions.
