+++ src/guardrails/prTerminal.ts
@@ -1,30 +1,11 @@
-import type { GuardrailState, HookInput } from "./types.js";
-
-export enum PrTerminalVerdict {
-  Merged = "merged",
-  Closed = "closed",
-}
-
-export interface PrTerminalEvent {
-  prNumber: number;
-  verdict: PrTerminalVerdict;
-}
-
-// gh success lines: `✓ Merged pull request #12 (title)` — also the
-// "Squashed and merged" / "Rebased and merged" strategy variants — and
-// `✓ Closed pull request owner/repo#12 (title)`. Newer gh prefixes the number
-// with owner/repo, older prints a bare `#12`; the optional [\w./-]* run accepts
-// both. Anchoring on the full verb phrase (never a bare MERGED/CLOSED token) is
-// what keeps `"state":"MERGED"` out: every `gh pr view --json`/API fetch
-// carries that token, and matching it would arm the gate on a routine status
-// query. It also excludes git's own merge-commit subject ("Merge pull
-// request #N"), which differs by verb tense.
-const GH_MERGED_LINE = /\b(?:Merged|Squashed and merged|Rebased and merged) pull request [\w./-]*#(\d+)/;
-const GH_CLOSED_LINE = /\bClosed pull request [\w./-]*#(\d+)/;
-// The pr-followup watch monitor's exit line (e.g. `TERMINAL pr=331: MERGED`),
-// which also surfaces when a monitor event notification is replayed through a
-// tool result.
-const MONITOR_TERMINAL_LINE = /\bTERMINAL pr=(\d+): (MERGED|CLOSED)\b/;
+import { PrTerminalGateAction, PrTerminalVerdict } from "./types.js";
+import type { GuardrailState, HookInput, PrTerminalEvent, PrTerminalGateDecision } from "./types.js";
+import {
+  GH_PR_MERGED_LINE,
+  GH_PR_CLOSED_LINE,
+  PR_MONITOR_TERMINAL_LINE,
+  MAX_PR_TERMINAL_BLOCKS,
+} from "./constants.js";
 
 // gh writes its success line to stderr, the monitor exit line arrives on
 // stdout/output, and a replayed Monitor notification may carry it in content —
@@ -35,15 +16,15 @@ export function detectPrTerminal(input: HookInput): PrTerminalEvent | null {
   const haystack = [response?.stdout, response?.stderr, response?.output, response?.content]
     .filter((part): part is string => typeof part === "string")
     .join("\n");
-  const mergedMatch = haystack.match(GH_MERGED_LINE);
+  const mergedMatch = haystack.match(GH_PR_MERGED_LINE);
   if (mergedMatch) {
     return { prNumber: Number(mergedMatch[1]), verdict: PrTerminalVerdict.Merged };
   }
-  const closedMatch = haystack.match(GH_CLOSED_LINE);
+  const closedMatch = haystack.match(GH_PR_CLOSED_LINE);
   if (closedMatch) {
     return { prNumber: Number(closedMatch[1]), verdict: PrTerminalVerdict.Closed };
   }
-  const monitorMatch = haystack.match(MONITOR_TERMINAL_LINE);
+  const monitorMatch = haystack.match(PR_MONITOR_TERMINAL_LINE);
   if (monitorMatch) {
     return {
       prNumber: Number(monitorMatch[1]),
@@ -79,19 +60,6 @@ export function applyNextOptionsOffered(state: GuardrailState): GuardrailState {
     terminalHandled: [...(state.terminalHandled ?? []), ...pending],
     terminalBlockCount: 0,
   };
-}
-
-export const MAX_PR_TERMINAL_BLOCKS = 3;
-
-export enum PrTerminalGateAction {
-  Block = "block",
-  Release = "release",
-  None = "none",
-}
-
-export interface PrTerminalGateDecision {
-  action: PrTerminalGateAction;
-  blockCount: number;
 }
 
 // Decide what the Stop hook does when a terminal PR's handoff is still owed.
