+++ src/guardrails/types.ts
@@ -11,6 +11,27 @@ export interface GuardrailState {
   terminalBlockCount?: number;
 }
 
+export enum PrTerminalVerdict {
+  Merged = "merged",
+  Closed = "closed",
+}
+
+export interface PrTerminalEvent {
+  prNumber: number;
+  verdict: PrTerminalVerdict;
+}
+
+export enum PrTerminalGateAction {
+  Block = "block",
+  Release = "release",
+  None = "none",
+}
+
+export interface PrTerminalGateDecision {
+  action: PrTerminalGateAction;
+  blockCount: number;
+}
+
 export interface HookInput {
   session_id?: string;
   cwd?: string;
